# docs/api/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Seismik Platform API",
    "version": "0.4.0"
  },
  "paths": {
...
    "/v1/devices/register": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Register Device",
        "operationId": "register_device_v1_devices_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegistration"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceRegistrationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/devices/unregister": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "Unregister Device",
        "operationId": "unregister_device_v1_devices_unregister_post",
        "parameters": [
          {
            "name": "X-Seismik-Device-Session",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Seismik-Device-Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUnregister"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "boolean"
                  },
                  "title": "Response Unregister Device V1 Devices Unregister Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/developer/webhooks": {
      "get": {
        "tags": [
          "organization-webhooks"
        ],
        "summary": "List Webhooks",
        "operationId": "list_webhooks_v1_developer_webhooks_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "seismik_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seismik Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "organization-webhooks"
        ],
        "summary": "Create Webhook",
        "operationId": "create_webhook_v1_developer_webhooks_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "seismik_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seismik Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WebhookCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookSecretResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/developer/webhooks/{webhook_id}": {
      "delete": {
        "tags": [
          "organization-webhooks"
        ],
        "summary": "Disable Webhook",
        "operationId": "disable_webhook_v1_developer_webhooks__webhook_id__delete",
        "parameters": [
          {
            "name": "webhook_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Webhook Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "seismik_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seismik Session"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceRegistration": {
        "properties": {
          "device_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "title": "Device Id"
          },
          "platform": {
            "$ref": "#/components/schemas/Platform"
          },
          "apns_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512,
                "minLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Apns Token"
          },
          "fcm_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4096,
                "minLength": 32
              },
              {
                "type": "null"
              }
            ],
            "title": "Fcm Token"
          },
          "country_code": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Country Code"
          },
          "zone_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Zone Id"
          },
          "latitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 90.0,
                "minimum": -90.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Latitude"
          },
          "longitude": {
            "anyOf": [
              {
                "type": "number",
                "maximum": 180.0,
                "minimum": -180.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Longitude"
          },
          "critical_alerts_authorized": {
            "type": "boolean",
            "title": "Critical Alerts Authorized",
            "default": false
          },
          "receive_early_alerts": {
            "type": "boolean",
            "title": "Receive Early Alerts",
            "default": true
          },
          "receive_official_updates": {
            "type": "boolean",
            "title": "Receive Official Updates",
            "default": true
          },
          "minimum_notification_magnitude": {
            "type": "number",
            "maximum": 10.0,
            "minimum": 0.0,
            "title": "Minimum Notification Magnitude",
            "default": 4.0
          },
          "alert_radius_km": {
            "type": "number",
            "maximum": 2000.0,
            "minimum": 10.0,
            "title": "Alert Radius Km",
            "description": "Radio máximo, elegido por la persona, para recibir avisos",
            "default": 250.0
          },
          "locale": {
            "type": "string",
            "maxLength": 16,
            "minLength": 2,
            "title": "Locale",
            "default": "es"
          },
          "app_attest_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8192,
                "minLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "App Attest Token",
            "description": "Firebase App Check token backed by App Attest/DeviceCheck"
          },
          "play_integrity_token": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 8192,
                "minLength": 16
              },
              {
                "type": "null"
              }
            ],
            "title": "Play Integrity Token",
            "description": "Firebase App Check token backed by Play Integrity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_id",
          "platform"
        ],
        "title": "DeviceRegistration",
        "x-product": "Seismik"
      },
      "DeviceRegistrationResponse": {
        "properties": {
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "registered": {
            "type": "boolean",
            "title": "Registered"
          },
          "crowd_token": {
            "type": "string",
            "title": "Crowd Token"
          },
          "device_session_token": {
            "type": "string",
            "title": "Device Session Token"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_id",
          "registered",
          "crowd_token",
          "device_session_token"
        ],
        "title": "DeviceRegistrationResponse",
        "x-product": "Seismik"
      },
      "DeviceUnregister": {
        "properties": {
          "device_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 8,
            "title": "Device Id"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "device_id"
        ],
        "title": "DeviceUnregister",
        "x-product": "Seismik"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Platform": {
        "type": "string",
        "enum": [
          "ios",
          "android"
        ],
        "title": "Platform"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WebhookCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 60,
            "minLength": 3,
            "title": "Name"
          },
          "endpoint": {
            "type": "string",
            "maxLength": 2083,
            "minLength": 1,
            "format": "uri",
            "title": "Endpoint"
          },
          "event_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "uniqueItems": true,
            "title": "Event Types"
          },
          "mode": {
            "type": "string",
            "const": "simulation_only",
            "title": "Mode",
            "default": "simulation_only"
          }
        },
        "type": "object",
        "required": [
          "name",
          "endpoint"
        ],
        "title": "WebhookCreateRequest"
      },
      "WebhookListResponse": {
        "properties": {
          "webhooks": {
            "items": {
              "$ref": "#/components/schemas/WebhookSummary"
            },
            "type": "array",
            "title": "Webhooks"
          },
          "safety_notice": {
            "type": "string",
            "title": "Safety Notice",
            "default": "Integración de simulación: nunca conecta ni autoriza controles físicos de IoT."
          }
        },
        "type": "object",
        "required": [
          "webhooks"
        ],
        "title": "WebhookListResponse"
      },
      "WebhookSecretResponse": {
        "properties": {
          "webhook_id": {
            "type": "string",
            "title": "Webhook Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "event_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Event Types"
          },
          "mode": {
            "type": "string",
            "const": "simulation_only",
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "last_delivered_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Delivered At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          },
          "signing_secret": {
            "type": "string",
            "title": "Signing Secret"
          },
          "warning": {
            "type": "string",
            "title": "Warning",
            "default": "Guárdalo ahora: Seismik no volverá a mostrar este secreto."
          }
        },
        "type": "object",
        "required": [
          "webhook_id",
          "name",
          "endpoint",
          "event_types",
          "mode",
          "status",
          "created_at",
          "signing_secret"
        ],
        "title": "WebhookSecretResponse"
      },
      "WebhookSummary": {
        "properties": {
          "webhook_id": {
            "type": "string",
            "title": "Webhook Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "event_types": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Event Types"
          },
          "mode": {
            "type": "string",
            "const": "simulation_only",
            "title": "Mode"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "disabled"
            ],
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "last_delivered_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Delivered At"
          },
          "last_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Error"
          }
        },
        "type": "object",
        "required": [
          "webhook_id",
          "name",
          "endpoint",
          "event_types",
          "mode",
          "status",
          "created_at"
        ],
        "title": "WebhookSummary"
      }
    }
  }
}